memBook.js > inputAjax(searchData)

+++ src/main/webapp/resources/js/empNewBook.js
@@ -1,5 +1,6 @@
 $(document).ready(function() {
-	$('#book').hide();
+	$('#bookGridDiv').hide();
+	$('#searchGrid').hide();
 	
 	$('#addRowBtn').click(function() {
 		addRowData();
@@ -16,6 +17,13 @@ $(document).ready(function() {
 	$('#inputBookBtn').click(function() {
 		inputBookAjax();
 	});
+	$('#searchBtn').click(function() {
+		$('#searchGrid').show();
+		testBookApi();
+	});
+	$('#searchInputBtn').click(function() {
+		searchInputAjax();
+	});
 	showBookGrid();
 	
 });
@@ -31,12 +39,73 @@ function showBookGrid() {
 		}
 		console.log('flag = ', flag);
 		if(flag == 1) {
-			$('#book').show();
-				console.log(grid.getData()[0]);
-				console.log(grid.getData()[0].new_status);
-				bookGrid.checkAll();
+			$('#bookGridDiv').show();
+			console.log(grid.getData()[0]);
+			console.log(grid.getData()[0].new_status);
+			bookGrid.checkAll();
 		}
 	}, 300);  
+}
+
+function testBookApi() {
+	$.ajax({
+		url: "https://dapi.kakao.com/v3/search/book",
+		headers: {'Authorization': 'KakaoAK 3e527a0c575e552fee7c82cf676cf81f'},
+		type: "get",
+		data: {
+			query: document.getElementById("search").value,
+			size: 50
+		},
+		success: function(result) {
+			console.log(result);
+			for (var i = 0; i < result.documents.length; i++) {
+				result.documents[i].thumbnail = '<img alt="no thumbnail" src="'+result.documents[i].thumbnail+'">';
+			}
+			searchGrid.uncheckAll();
+			searchGrid.resetData(result.documents);
+		}
+	})
+}
+
+function searchInputAjax() {
+	var length = searchGrid.getCheckedRows().length;
+	
+	if(length != 0) {
+		console.log(searchGrid.getCheckedRows());
+		for(var i = 0; i < length; i++) {
+			searchData(searchGrid.getCheckedRows()[i]);
+		}
+		$('#search').val("");
+		$('#searchGrid').hide();
+	} else {
+		alert("신청 추가할 도서를 선택해주세요.");
+	}
+}
+
+function searchData(array) {
+	var author = "";
+	
+	var option = {
+			at:0,
+			focus:true
+	};
+	
+	if(array.authors.length != 1) {
+		for (var i = 0; i < array.authors.length - 1; i++) {
+			author += array.authors[i] + ", ";
+		}
+	}
+	author += array.authors[array.authors.length - 1];
+	console.log(author);
+	
+	var data = {
+			book_name: array.title, book_author: author, book_pub_house: array.publisher,
+			book_pub_date: array.datetime.substring(0, 10), book_ISBN: array.isbn, book_price: array.price 
+	}
+	
+	console.log(data);
+	grid.appendRow(data,option);
+	grid.check(grid.getRowAt(0).rowKey);
 }
 
 function addRowData() {
@@ -55,6 +124,8 @@ function addRowData() {
 
 function inputAjax() {
 //	console.log('focus success >> ' + grid.focus(grid.getRowAt(0).rowKey, 'book_name', true));
+	console.log("input() > ", grid.getCheckedRows());
+	
 	if(grid.getCheckedRows().length != 0) {
 		var checkedGrid= grid.getCheckedRows();
 		console.log(checkedGrid);
@@ -152,21 +223,23 @@ function updateAjax() {
 			alert("승인한 항목은 수정이 불가능합니다. \n다시 선택한 후 수정해주세요.");
 		} else {
 			grid.focus(grid.getRowAt(0).rowKey, 'new_book_num', true);
-			$.ajax({
-				type : "POST",
-				contentType : "application/json;charset=UTF-8",
-				dataType : "json",
-				data : JSON.stringify(grid.getCheckedRows()),
-				url : "./newApply/update.do",
-				success : function(data){
-					console.log(data.result);
-					grid.uncheckAll();
-					confirm();
-				},
-				error : function(e) {
-					alert('Error : ' + e);
-				}
-			});
+			setTimeout(function() {
+				$.ajax({
+					type : "POST",
+					contentType : "application/json;charset=UTF-8",
+					dataType : "json",
+					data : JSON.stringify(grid.getCheckedRows()),
+					url : "./newApply/update.do",
+					success : function(data){
+						console.log(data.result);
+						grid.uncheckAll();
+						confirm();
+					},
+					error : function(e) {
+						alert('Error : ' + e);
+					}
+				});
+			}, 200); 
 		}
 	} else {
 		alert("수정할 도서를 선택해주세요.");
@@ -174,30 +247,43 @@ function updateAjax() {
 }
 
 function inputBookAjax() {
-	if(bookGrid.getCheckedRows().length != 0) {
-		console.log(bookGrid.getCheckedRows());
-		bookGrid.focus(bookGrid.getRowAt(0).rowKey, 'book_num', true);
-		$.ajax({
-			type : "POST",
-			contentType : "application/json;charset=UTF-8",
-			dataType : "json",
-			data : JSON.stringify(bookGrid.getCheckedRows()),
-			url : "./newApply/inputBook.do",
-			success : function(data){
-				console.log(data.result);
-				confirm();
-			},
-			error : function(e) {
-				alert('Error : ' + e);
+	bookGrid.focus(bookGrid.getRowAt(0).rowKey, 'book_name', true);
+	setTimeout(function() {
+		if(bookGrid.getCheckedRows().length != 0) {
+			console.log(bookGrid.getCheckedRows());
+			var flag = 0;
+			var checkedGrid= bookGrid.getCheckedRows();
+			for (var i = 0; i < checkedGrid.length; i++) {
+				console.log(i, " checkedGrid[i].book_ctgr_num >> ", checkedGrid[i].book_ctgr_num);
+				if(checkedGrid[i].book_ctgr_num == '입력' || checkedGrid[i].book_ctgr_num <= 0 || checkedGrid[i].book_ctgr_num >= 1000) 
+					flag = 1;
 			}
-		});
-		
-		grid.checkAll();
-		deleteNewBookAjax();
-		console.log("complete");
-	} else {
-		alert("입력할 도서를 선택해주세요.");
-	}
+			console.log("inputBook() > ", flag);
+			if(flag == 0) {
+					$.ajax({
+						type : "POST",
+						contentType : "application/json;charset=UTF-8",
+						dataType : "json",
+						data : JSON.stringify(bookGrid.getCheckedRows()),
+						url : "./newApply/inputBook.do",
+						success : function(data){
+							console.log(data.result);
+							confirm();
+							grid.checkAll();
+							deleteNewBookAjax();
+							console.log("complete");
+						},
+						error : function(e) {
+							alert('Error : ' + e);
+						}
+					});
+			} else {
+				alert("분류기호를 1.0~999.9 사이의 숫자로 입력해주세요.")
+			}
+		} else {
+			alert("입력할 도서를 선택해주세요.");
+		}
+	}, 200);  
 }
 
 function confirm(){
@@ -220,6 +306,48 @@ var gridData2 =
 			readData: { url: 'http://localhost:8080/mvc/book/newApply.do/readData2', method: 'GET' }
 	}
 }
+
+const searchGrid = new tui.Grid({
+	el: document.getElementById('searchGrid'),
+	data: null,
+	bodyHeight: 360,
+	rowHeight: 180,
+	rowHeaders: ['rowNum','checkbox'],
+	columns: [
+		{
+			header: '이미지',
+			name: 'thumbnail'
+		},
+		{
+			header: '제목',
+			name: 'title'
+		},
+		{
+			header: '저자',
+			name: 'authors'
+		},
+		{
+			header: '출판사',
+			name: 'publisher'
+		},
+		{
+			header: '발행일',
+			name: 'datetime'
+		},
+		{
+			header: 'isbn',
+			name: 'isbn'
+		},
+		{
+			header: '정가',
+			name: 'price'
+		}
+	],
+	columnOptions: {
+	      resizable: true
+	}
+});
+
 const grid = new tui.Grid({
 	el: document.getElementById('grid'),
 	data: gridData,
@@ -235,6 +363,10 @@ const grid = new tui.Grid({
 			name: 'new_book_num'
 		},
 		{
+			header: '신청 도서 번호',
+			name: 'new_book_num_sub'
+		},
+		{
 			header: '도서명',
 			name: 'book_name',
 			editor: 'text'
@@ -247,6 +379,16 @@ const grid = new tui.Grid({
 		{
 			header: '출판사명',
 			name: 'book_pub_house',
+			editor: 'text'
+		},
+		{
+			header: '발행일',
+			name: 'book_pub_date',
+			editor: 'datePicker'
+		},
+		{
+			header: 'ISBN',
+			name: 'book_ISBN',
 			editor: 'text'
 		},
 		{
@@ -279,6 +421,9 @@ const grid = new tui.Grid({
 			}
 		},
 		
+	},
+	columnOptions: {
+	      resizable: true
 	}
 });
 
@@ -314,15 +459,9 @@ const bookGrid = new tui.Grid({
 			editor: 'text'
 		},
 		{
-			header: '발행년도',
+			header: '발행일',
 			name: 'book_pub_date',
-			editor: {
-				type: 'datePicker',
-				options: {
-					format: 'yyyy',
-					type: 'year'
-				}
-			}
+			editor: 'datePicker'
 		},
 		{
 			header: 'ISBN',
@@ -349,7 +488,7 @@ const bookGrid = new tui.Grid({
 		},
 		{
 			header: '대여여부',
-			name: 'rent',
+			name: 'rent'/*,
 			editor: {
 				type: 'radio',
 				options: {
@@ -358,11 +497,11 @@ const bookGrid = new tui.Grid({
 						{ text: '대여중', value: '대여중' },
 					]
 				}
-			}
+			}*/
 		},
 		{
 			header: '예약여부',
-			name: 'book_rsrv_status',
+			name: 'book_rsrv_status'/*,
 			editor: {
 				type: 'radio',
 				options: {
@@ -371,12 +510,15 @@ const bookGrid = new tui.Grid({
 						{ text: '예약중', value: '예약중' },
 					]
 				}
-			}
+			}*/
 		},
 		{
 			header: '입력일',
 			name: 'book_input_date',
 			
 		}
-	]
+	],
+	columnOptions: {
+	      resizable: true
+	}
 });
